Allow gencorpora to replay events
